requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
gdown==5.2.0
google-generativeai==0.8.4
python-dotenv==1.0.1
requests==2.32.3
